requirements.txt
asgiref==3.10.0
Django==5.2.8
sqlparse==0.5.3
tzdata==2025.2
django-import-export==4.3.0
openpyxl==3.1.2
